GRP-931: Rule : problem with ruleThenEnum assignMembershipDisabledDaysForOwnerGroupId wraps integer limit needs long
git-svn-id: https://svn.internet2.edu/svn/i2mi/branches/GROUPER_2_1_BRANCH@8731 bc24b658-cf39-4249-b827-e3526eaaeb42

+++ grouper-misc/grouperClient/src/java/edu/internet2/middleware/grouperClient/discovery/DiscoveryClient.java
@@ -804,7 +804,7 @@ public class DiscoveryClient {
         Date date = dateFormat.parse(datePart);
 
         //see if more than 20 minutes old
-        if (System.currentTimeMillis() - date.getTime() < 20 * 60 * 1000) {
+        if (System.currentTimeMillis() - date.getTime() < (long)20 * 60 * 1000) {
           continue;
         }
